% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile.de_scraping.R
\name{mobile.de_scraping}
\alias{mobile.de_scraping}
\title{mobile.de_scraping}
\usage{
mobile.de_scraping(url, lang = "en")
}
\arguments{
\item{url}{Enter an url copied from mobile.de web}
}
\value{
A data frame with all items avalaible in mobile.de
}
\description{
mobile.de_scraping
}
\examples{
mobile.de_scraping(url)
}
